+++ src/riccati/sparse_ros2.jl
@@ -5,6 +5,7 @@ function _solve(
     alg::Ros2;
     dt::Real,
     save_state::Bool,
+    observer,
 ) where {TL,TD}
     @unpack E, A, B, C, tspan = prob
     q = size(C, 1)
@@ -23,6 +24,9 @@ function _solve(
     K = BᵀLD*(L'*E)
     Ks = [K]
     sizehint!(Ks, len)
+
+    observe_gdre_start!(observer, prob, Ros2())
+    observe_gdre_step!(observer, tstops[1], X, K)
 
     for i in 2:len
         τ = tstops[i-1] - tstops[i]
@@ -45,7 +49,7 @@ function _solve(
         R1 = compress!(LDLᵀ{TL,TD}(G, S))
 
         lyap = GALEProblem(E, F, R1)
-        K1 = solve(lyap, ADI())
+        K1 = solve(lyap, ADI(); observer)
 
         # Solve Lyapunov equation of 2nd stage
         T₁, D₁ = K1
@@ -55,7 +59,7 @@ function _solve(
         R2 = LDLᵀ{TL,TD}(G₂, S₂)
 
         lyap = GALEProblem(E, F, R2)
-        K2 = solve(lyap, ADI())
+        K2 = solve(lyap, ADI(); observer)
 
         # Update X
         X = X + ((2-1/2γ)*τ)*K1 + (-τ/2)*K2
@@ -66,8 +70,12 @@ function _solve(
         BᵀLD = (B'*L)*D
         K = BᵀLD*(L'*E)
         push!(Ks, K)
+
+        observe_gdre_step!(observer, tstops[i], X, K)
     end
     save_state || push!(Xs, X)
+
+    observe_gdre_done!(observer)
 
     return DRESolution(Xs, Ks, tstops)
 end