+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # DEV
 
 * Fix LDLᵀ compression for indefinite objects (d6a649ad62ab6d413ce82e5a2b0090de813a33de)
+* Add callbacks to allow user to gather information during `solve` calls;
+  see docstring of the `Callbacks` module for more info
 
 # v0.3